+++ poker_eval_libraries.rb
@@ -1,0 +1,48 @@
+module PokerEvalLibraries
+
+  def library_of_values
+    {
+      'A': 'Ace',
+      'K': 'King',
+      'Q': 'Queen',
+      'J': 'Jack',
+      '10': 'Ten',
+      '9': 'Nine',
+      '8': 'Eight',
+      '7': 'Seven',
+      '6': 'Six',
+      '5': 'Five',
+      '4': 'Four',
+      '3': 'Three',
+      '2': 'Two'
+    }
+  end
+
+  def library_of_suites
+    {
+      'd': 'diamonds',
+      'h': 'hearts',
+      's': 'spades',
+      'c': 'clubs'
+    }
+  end
+
+  def numeric_values
+    {
+      'A': 14,
+      'K': 13,
+      'Q': 12,
+      'J': 11,
+      '10': 10,
+      '9': 9,
+      '8': 8,
+      '7': 7,
+      '6': 6,
+      '5': 5,
+      '4': 4,
+      '3': 3,
+      '2': 2,
+    }
+  end
+
+end
